requirements.txt
fuzzywuzzy==0.18.0
Levenshtein==0.25.1
pillow==10.3.0
python-Levenshtein==0.25.1
rapidfuzz==3.9.3
